Module poem::endpoint

Endpoint related types.

Structs

After

Endpoint for the after method.

AndThen

Endpoint for the and_then method.

Around

Endpoint for the around method.

Before

Endpoint for the before method.

CatchAllError

Endpoint for the catch_all_error method.

CatchError

Endpoint for the catch_error method.

InspectAllError

Endpoint for the inspect_all_err method.

InspectError

Endpoint for the inspect_err method.

Map

Endpoint for the map_ok method.

MapToResponse

Endpoint for the map_to_response method.

ToResponse

Endpoint for the to_response method.

Traits

Endpoint

An HTTP request handler.

EndpointExt

Extension trait for Endpoint.

IntoEndpoint

Represents a type that can convert into endpoint.

Functions

make

Create an endpoint with a asyncness function.

make_sync

Create an endpoint with a function.

Type Definitions

BoxEndpoint

An owned dynamically typed Endpoint for use in cases where you can’t statically type your result or need to add some indirection.