Module poem::error

Some common error types.

Structs

Error

General response error.

GetDataError

A possible error value when extracts data from request fails.

MethodNotAllowedError

Error occurred in the router.

MissingJsonContentTypeError

A missing json Content-Type error value when parsing header.

NotFoundError

Error occurred in the router.

ParseJsonError

A possible error value when parsing JSON.

ParsePathError

Only the endpoints under the router can get the path parameters, otherwise this error will occur.

ParseQueryError

A possible error value when parsing query.

Enums

CorsError

A possible error value occurred in the Cors middleware.

ParseCookieError

A possible error value when parsing cookie.

ParseFormError

A possible error value when parsing form.

ParseMultipartError

A possible error value when parsing multipart.

ParseTypedHeaderError

A possible error value when parsing typed headers.

ReadBodyError

A possible error value when reading the body.

RouteError

A possible error value occurred when adding a route.

SizedLimitError

A possible error value occurred in the SizeLimit middleware.

StaticFileError

A possible error value when processing static files.

UpgradeError

A possible error value when upgrading connection.

Traits

IntoResult

Represents a type that can be converted to poem::Result<T>.

ResponseError

Represents a type that can be converted to Error.

Functions

BadGateway

Wraps any error into Error and the status code is StatusCode::BAD_GATEWAY.

BadRequest

Wraps any error into Error and the status code is StatusCode::BAD_REQUEST.

Conflict

Wraps any error into Error and the status code is StatusCode::CONFLICT.

ExpectationFailed

Wraps any error into Error and the status code is StatusCode::EXPECTATION_FAILED.

FailedDependency

Wraps any error into Error and the status code is StatusCode::FAILED_DEPENDENCY.

Forbidden

Wraps any error into Error and the status code is StatusCode::FORBIDDEN.

GatewayTimeout

Wraps any error into Error and the status code is StatusCode::GATEWAY_TIMEOUT.

Gone

Wraps any error into Error and the status code is StatusCode::GONE.

HttpVersionNotSupported

Wraps any error into Error and the status code is StatusCode::HTTP_VERSION_NOT_SUPPORTED.

ImATeapot

Wraps any error into Error and the status code is StatusCode::IM_A_TEAPOT.

InsufficientStorage

Wraps any error into Error and the status code is StatusCode::INSUFFICIENT_STORAGE.

InternalServerError

Wraps any error into Error and the status code is StatusCode::INTERNAL_SERVER_ERROR.

LengthRequired

Wraps any error into Error and the status code is StatusCode::LENGTH_REQUIRED.

Locked

Wraps any error into Error and the status code is StatusCode::LOCKED.

LoopDetected

Wraps any error into Error and the status code is StatusCode::LOOP_DETECTED.

MethodNotAllowed

Wraps any error into Error and the status code is StatusCode::METHOD_NOT_ALLOWED.

MisdirectedRequest

Wraps any error into Error and the status code is StatusCode::MISDIRECTED_REQUEST.

NetworkAuthenticationRequired

Wraps any error into Error and the status code is StatusCode::NETWORK_AUTHENTICATION_REQUIRED.

NotAcceptable

Wraps any error into Error and the status code is StatusCode::NOT_ACCEPTABLE.

NotExtended

Wraps any error into Error and the status code is StatusCode::NOT_EXTENDED.

NotFound

Wraps any error into Error and the status code is StatusCode::NOT_FOUND.

NotImplemented

Wraps any error into Error and the status code is StatusCode::NOT_IMPLEMENTED.

PayloadTooLarge

Wraps any error into Error and the status code is StatusCode::PAYLOAD_TOO_LARGE.

PaymentRequired

Wraps any error into Error and the status code is StatusCode::PAYMENT_REQUIRED.

PreconditionFailed

Wraps any error into Error and the status code is StatusCode::PRECONDITION_FAILED.

PreconditionRequired

Wraps any error into Error and the status code is StatusCode::PRECONDITION_REQUIRED.

ProxyAuthenticationRequired

Wraps any error into Error and the status code is StatusCode::PROXY_AUTHENTICATION_REQUIRED.

RangeNotSatisfiable

Wraps any error into Error and the status code is StatusCode::RANGE_NOT_SATISFIABLE.

RequestHeaderFieldsTooLarge

Wraps any error into Error and the status code is StatusCode::REQUEST_HEADER_FIELDS_TOO_LARGE.

RequestTimeout

Wraps any error into Error and the status code is StatusCode::REQUEST_TIMEOUT.

ServiceUnavailable

Wraps any error into Error and the status code is StatusCode::SERVICE_UNAVAILABLE.

TooManyRequests

Wraps any error into Error and the status code is StatusCode::TOO_MANY_REQUESTS.

Unauthorized

Wraps any error into Error and the status code is StatusCode::UNAUTHORIZED.

UnavailableForLegalReasons

Wraps any error into Error and the status code is StatusCode::UNAVAILABLE_FOR_LEGAL_REASONS.

UnprocessableEntity

Wraps any error into Error and the status code is StatusCode::UNPROCESSABLE_ENTITY.

UnsupportedMediaType

Wraps any error into Error and the status code is StatusCode::UNSUPPORTED_MEDIA_TYPE.

UpgradeRequired

Wraps any error into Error and the status code is StatusCode::UPGRADE_REQUIRED.

UriTooLong

Wraps any error into Error and the status code is StatusCode::URI_TOO_LONG.

VariantAlsoNegotiates

Wraps any error into Error and the status code is StatusCode::VARIANT_ALSO_NEGOTIATES.

Type Definitions

Result

A specialized Result type for Poem.