Module poem::listener

Commonly used listeners.

Structs

BoxIo

An IO type for BoxAcceptor.

BoxListener

An owned dynamically typed Listener for use in cases where you can’t statically type your result or need to add some indirection.

Combined

Listener for the Listener::combine and AcceptorExt::combine method.

TcpAcceptor

A acceptor that accepts TCP connections.

TcpListener

A TCP listener.

UnixAcceptor

A acceptor that accepts connections.

UnixListener

A Unix domain socket listener.

Enums

CombinedStream

A IO stream for CombinedAcceptor.

Traits

Acceptor

Represents a acceptor type.

AcceptorExt

Extension trait for Acceptor.

Listener

Represents a listener that can be listens for incoming connections.

Type Definitions

BoxAcceptor

An owned dynamically typed Acceptor for use in cases where you can’t statically type your result or need to add some indirection.