Trait poem::listener::Listener

pub trait Listener: Send {
    type Acceptor: Acceptor;

    fn into_acceptor<'async_trait>(
        self
    ) -> Pin<Box<dyn Future<Output = IoResult<Self::Acceptor>> + Send + 'async_trait>>
    where
        Self: 'async_trait;

    fn combine<T>(self, other: T) -> Combined<Self, T>
    where
        Self: Sized,
    { ... }
    fn boxed(self) -> BoxListener
    where
        Self: Sized + 'static,
    { ... }
}

Represents a listener that can be listens for incoming connections.

Required Associated Types

type Acceptor: Acceptor

The acceptor type.

Required Methods

fn into_acceptor<'async_trait>(
self
) -> Pin<Box<dyn Future<Output = IoResult<Self::Acceptor>> + Send + 'async_trait>> where
Self: 'async_trait,

Create a acceptor instance.

Provided Methods

fn combine<T>(self, other: T) -> Combined<Self, T> where
Self: Sized,

Combine two listeners.

You can call this function multiple times to combine more listeners.

Example

use poem::listener::{Listener, TcpListener};

let listener = TcpListener::bind("127.0.0.1:80").combine(TcpListener::bind("127.0.0.1:81"));

fn boxed(self) -> BoxListener where
Self: Sized + 'static,

Wrap the listener in a Box.