Module poem::middleware

Commonly used middleware.

Structs

AddData

Middleware for add any data to request.

AddDataEndpoint

Endpoint for AddData middleware.

CatchPanic

Middleware for catches panics and converts them into 500 INTERNAL SERVER ERROR responses.

CatchPanicEndpoint

Endpoint for PanicHandler middleware.

CookieJarManager

Middleware for CookieJar support.

CookieJarManagerEndpoint

Endpoint for CookieJarManager middleware.

Cors

Middleware for CORS

CorsEndpoint

Endpoint for Cors middleware.

FnMiddleware

A middleware implemented by a closure.

ForceHttps

Middleware for force redirect to HTTPS uri.

NormalizePath

Middleware for normalizing a request’s path so that routes can be matched more flexibly.

NormalizePathEndpoint

Endpoint for NormalizePath middleware.

PropagateHeader

Middleware for propagate a header from the request to the response.

PropagateHeaderEndpoint

Endpoint for PropagateHeader middleware.

SensitiveHeader

Middleware for mark headers value represents sensitive information.

SensitiveHeaderEndpoint

Endpoint for SensitiveHeader middleware.

SetHeader

Middleware for override/append headers to response.

SetHeaderEndpoint

Endpoint for SetHeader middleware.

SizeLimit

Middleware for limit the request payload size.

SizeLimitEndpoint

Endpoint for SizeLimit middleware.

Tracing

Middleware for tracing.

TracingEndpoint

Endpoint for Tracing middleware.

Enums

TrailingSlash

Determines the behavior of the NormalizePath middleware.

Traits

Middleware

Represents a middleware trait.

PanicHandler

Panics handler

Functions

make

Make middleware with a closure.