Struct poem::Body

pub struct Body(_);

A body object for requests and responses.

Implementations

impl Body

pub fn from_bytes(data: Bytes) -> Self

Create a body object from Bytes.

pub fn from_string(data: String) -> Self

Create a body object from String.

pub fn from_vec(data: Vec<u8>) -> Self

Create a body object from Vec<u8>.

pub fn from_async_read(reader: impl AsyncRead + Send + 'static) -> Self

Create a body object from reader.

pub fn from_bytes_stream<S, O, E>(stream: S) -> Self where
S: Stream<Item = Result<O, E>> + Send + 'static,
O: Into<Bytes> + 'static,
E: Error + Send + Sync + 'static,

Create a body object from bytes stream.

pub fn from_json(body: impl Serialize) -> Result<Self>

Create a body object from JSON.

pub fn empty() -> Self

Create an empty body.

pub fn is_empty(&self) -> bool

Returns true if this body is empty.

pub async fn into_bytes(self) -> Result<Bytes, ReadBodyError>

Consumes this body object to return a Bytes that contains all data.

pub async fn into_vec(self) -> Result<Vec<u8>, ReadBodyError>

Consumes this body object to return a Vec<u8> that contains all data.

pub async fn into_bytes_limit(
self,
limit: usize
) -> Result<Bytes, ReadBodyError>

Consumes this body object to return a Bytes that contains all data, returns Err(ReadBodyError::PayloadTooLarge) if the length of the payload exceeds limit.

Example

use poem::{error::ReadBodyError, handler, http::StatusCode, Body, Endpoint, Request, Result};

#[handler]
async fn index(data: Body) -> Result<()> {
    Ok(data.into_bytes_limit(5).await.map(|_| ())?)
}

let req = Request::builder().body("12345");
assert_eq!(index.get_response(req).await.status(), StatusCode::OK);

let req = Request::builder().body("123456");
assert_eq!(
    index.get_response(req).await.status(),
    StatusCode::PAYLOAD_TOO_LARGE
);

pub async fn into_string(self) -> Result<String, ReadBodyError>

Consumes this body object to return a String that contains all data.

pub async fn into_json<T: DeserializeOwned>(self) -> Result<T>

Consumes this body object and parse it as T.

Errors

ReadBodyError

ParseJsonError

pub fn into_async_read(self) -> impl AsyncRead + Unpin + Send + 'static

Consumes this body object to return a reader.

pub fn into_bytes_stream(
self
) -> impl Stream<Item = Result<Bytes, IoError>> + Send + 'static

Consumes this body object to return a bytes stream.