Struct poem::Response

pub struct Response { /* private fields */ }

Represents an HTTP response.

Implementations

impl Response

pub fn from_parts(parts: ResponseParts, body: Body) -> Self

Creates a new Response with the given head and body.

pub fn builder() -> ResponseBuilder

Creates a response builder.

pub fn status(&self) -> StatusCode

Returns the associated status code.

pub fn is_ok(&self) -> bool

Returns true if status code is StatusCode::OK.

pub fn is_success(&self) -> bool

Check if status is within 200-299.

pub fn set_status(&mut self, status: StatusCode)

Sets the status code for this response.

pub fn content_type(&self) -> Option<&str>

Returns the content type of this response.

pub fn set_content_type(self, content_type: impl AsRef<str>) -> Self

Sets the Content-Type header on the response.

pub fn headers(&self) -> &HeaderMap

Returns a reference to the associated header map.

pub fn headers_mut(&mut self) -> &mut HeaderMap

Returns a mutable reference to the associated header map.

pub fn header(&self, name: impl AsRef<str>) -> Option<&str>

Returns the string value of the specified header.

NOTE: Returns None if the header value is not a valid UTF8 string.

pub fn version(&self) -> Version

Returns the associated version.

pub fn set_version(&mut self, version: Version)

Sets the version for this response.

pub fn extensions(&self) -> &Extensions

Returns a reference to the associated extensions.

pub fn extensions_mut(&mut self) -> &mut Extensions

Returns a mutable reference to the associated extensions.

pub fn data<T: Send + Sync + 'static>(&self) -> Option<&T>

Get a reference from extensions, similar to self.extensions().get().

pub fn set_data(&mut self, data: impl Send + Sync + 'static)

Inserts a value to extensions, similar to self.extensions().insert(data).

pub fn set_body(&mut self, body: impl Into<Body>)

Sets the body for this response.

pub fn take_body(&mut self) -> Body

Take the body from this response and sets the body to empty.

pub fn into_body(self) -> Body

Consume this response and return its inner body.

pub fn into_parts(self) -> (ResponseParts, Body)

Consumes the response returning the head and body parts.