Module poem::web

Commonly used as the type of extractor or response.

Modules

cookie

Cookie related types.

headers

Typed HTTP Headers

sse

Server-Sent Events (SSE) types.

Structs

Accept

Accept header, defined in RFC7231

Data

An extractor that can extract data from the request extension.

Field

A single field in a multipart stream.

Form

An extractor that can deserialize some type from query string or body.

Html

An HTML response.

Json

JSON extractor and response.

LocalAddr

Local server’s address.

Multipart

An extractor that parses multipart/form-data requests commonly used with file uploads.

Path

An extractor that will get captures from the URL and parse them using serde.

Query

An extractor that can deserialize some type from query string.

RealIp

An extractor that can extracts the real ip from request headers

Redirect

A redirect response.

RemoteAddr

Remote peer’s address.

RequestBody

The body parameter type of FromRequest::from_request method.

TempFile

An extractor that extracts the body and writes the contents to a temporary file.

TypedHeader

An extractor that extracts a typed header value.

WithBody

Returned by with_body method.

WithContentType

Returned by with_content_type method.

WithHeader

Returned by with_header method.

WithStatus

Returned by with_header method.

Traits

FromRequest

Represents an type that can be extract from requests.

IntoResponse

Represents a type that can convert into response.