Module poem::web::cookie

Cookie related types.

Structs

Cookie

HTTP cookie extractor.

CookieJar

A collection of cookies that tracks its modifications.

PrivateCookieJar

A child cookie jar that provides authenticated encryption for its cookies.

SignedCookieJar

A child cookie jar that authenticates its cookies.

Type Definitions

CookieKey

A cryptographic master key for use with Signed and/or Private jars.

SameSite

The SameSite cookie attribute.